Updates for Patrick's code review
Mostly just converts ImageUtils static methods to Bitmap extensions
