+++ channel-sdk/src/main/java/com/zello/channel/sdk/image/Dimensions.kt
@@ -1,0 +1,7 @@
+package com.zello.channel.sdk.image
+
+internal data class Dimensions(val width: Int, val height: Int) {
+	companion object {
+		fun square(dimension: Int) = Dimensions(dimension, dimension)
+	}
+}
