--- channel-sdk/src/main/java/com/zello/channel/sdk/image/ImageUtils.kt
@@ -1,84 +1,0 @@
-package com.zello.channel.sdk.image
-
-import android.graphics.Bitmap
-import java.io.ByteArrayOutputStream
-
-/// The dimensions of the bitmap. Packages the width and height into a single object.
-internal val Bitmap.dimensions: Dimensions
-	get() {
-		return Dimensions(width, height)
-	}
-
-internal data class Dimensions(val width: Int, val height: Int) {
-	companion object {
-		fun square(dimension: Int) = Dimensions(dimension, dimension)
-	}
-}
-
-internal object ImageUtils {
-	/// This method should NOT be called on the main thread
-	fun bitmapWithMaxDimensions(bitmap: Bitmap, maxWidth: Int, maxHeight: Int): Bitmap {
-		if (bitmap.width <= maxWidth && bitmap.height <= maxHeight) {
-			return bitmap
-		}
-		val targetDimensions = dimensionsFitting(bitmap.dimensions, Dimensions(maxWidth, maxHeight))
-		return scaleBitmap(bitmap, targetDimensions.width, targetDimensions.height)
-	}
-
-	/**
-	 * Compresses a bitmap to JPEG with a maximum data length. Makes multiple compression passes at
-	 * reduced quality until it finds a JPEG quality that produces compressed data less than maxSize
-	 * bytes. If the image at quality 10 is still larger than maxSize, it uses that data anyway.
-	 *
-	 * @param image the image to compress
-	 * @param maxSize the preferred maximum data length in bytes
-	 * @return a ByteArray containing the compressed image data
-	 */
-	fun compressImage(image: Bitmap, maxSize: Int): ByteArray {
-		var dataLength: Int
-		var quality = 90
-		val minQuality = 10
-		var stream: ByteArrayOutputStream
-		do {
-			stream = ByteArrayOutputStream()
-			image.compress(Bitmap.CompressFormat.JPEG, quality, stream)
-			dataLength = stream.size()
-			// If image data is too large, reduce quality by 10 and try again
-			if (dataLength > maxSize) {
-				quality -= 10
-			}
-		} while (dataLength > maxSize && quality >= minQuality)
-		return stream.toByteArray() // Stream guaranteed to be non-null
-	}
-
-	private fun dimensionsFitting(original: Dimensions, max: Dimensions): Dimensions {
-		val widthScalingFactor = max.width.toFloat() / original.width.toFloat()
-		val heightScalingFactor = max.height.toFloat() / original.height.toFloat()
-		val targetHeight: Int
-		val targetWidth: Int
-		if (widthScalingFactor < heightScalingFactor) {
-			targetWidth = max.width
-			targetHeight = (original.height * widthScalingFactor).toInt()
-		} else {
-			targetHeight = max.height
-			targetWidth = (original.width * heightScalingFactor).toInt()
-		}
-		return Dimensions(targetWidth, targetHeight)
-	}
-
-	private fun scaleBitmap(bitmap: Bitmap, targetWidth: Int, targetHeight: Int): Bitmap {
-		var tempBitmap = bitmap
-		while (tempBitmap.width > targetWidth || tempBitmap.height > targetHeight) {
-			var passTargetWidth = targetWidth
-			var passTargetHeight = targetHeight
-			// If we need to shrink the image by more than 2x, make multiple passes to reduce aliasing
-			if ((targetWidth < tempBitmap.width / 2) || (targetHeight < tempBitmap.height / 2)) {
-				passTargetWidth = tempBitmap.width / 2
-				passTargetHeight = tempBitmap.height / 2
-			}
-			tempBitmap = Bitmap.createScaledBitmap(tempBitmap, passTargetWidth, passTargetHeight, true)
-		}
-
-		return tempBitmap
-	}
-}
